scripts/pyproject.toml
[project]
name = "newsfrontier-scripts"
version = "0.1.0"
description = "Development and management scripts for NewsFrontier"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.7",
    "passlib[bcrypt]>=1.7.4",
    "requests>=2.31.0",
    "click>=8.1.0",
    "colorama>=0.4.6",
    "python-dateutil>=2.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]